requirements.txt
transformers
streamlit
fuzzywuzzy
pandas
PyTorch
TensorFlow
tf-keras
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
SpeechRecognition
opencv-python
python-dotenv
numpy
pandas